moves global usable code to seperated global classes
